Add rails-better-views-on-ruby to illustrate foreground/background idea
The rails-better-views (and rails-short-sandbox before it) use a
foreground script that leverages "sh -x" output to give the user a sense
of progress while informing whem what commands are being run by the
background script.  It is built on ruby:3.2.2, but I'm about to switch
it to rails:7.1 which has most of these commands already run.

Copy the lab to rails-better-views-on-ruby to preserve the
forground/background example before removing it in rails-better-views.